Web_App/requirements.txt

flask
keras-ocr
transformers
evaluate

Pillow
pandas
matplotlib
numpy
werkzeug
